// src/app/playground/archived/radial-blur/config/types.ts
export type SemanticBgColor =
  | 'primary'
  | 'secondary'
  | 'tertiary'
  | 'quaternary'
  | 'brand-primary'
  | 'brand-secondary'

export type ShineStyle =
  | 'none'
  | 'shine-0'
  | 'shine-0-subtle'
  | 'shine-1'
  | 'shine-1-subtle'
  | 'shine-2'
  | 'shine-2-subtle'
  | 'shine-3'
  | 'shine-3-subtle'
  | 'shine-brand'
  | 'shine-brand-subtle'